+++ src/risk.rs
@@ -6,9 +6,12 @@ use rust_decimal_macros::dec;
 use crate::portfolio::{Portfolio, PositionHandlers};
 use crate::types::{Candle, Trade};
 
+#[derive(Debug)]
 pub enum RiskCalculationErrors {
     /// The market data and historical data are not aligned by timestamp
-    CandleDataNotAligned
+    CandleDataNotAligned,
+
+    CouldNotExtract
 }
 
 /// Calculate risk metrics for a portfolio against market data, and historical data for the asset.
@@ -38,7 +41,12 @@ pub fn calculate_risk(portfolio: &Portfolio, market_data: &[Candle], historical_
     let (total_position_value, average_entry_price, unrealized_pnl) = calculate_position_metrics(portfolio, current_price);
     let returns = calculate_returns(historical_data);
 
-    let value_at_risk = calculate_value_at_risk(&returns, total_position_value);
+    let value_at_risk = if total_position_value == Decimal::ZERO {
+        // Use historical data for initial VaR calculation
+        calculate_value_at_risk(&returns, Decimal::from(1_000)) // Assume a default position value
+    } else {
+        calculate_value_at_risk(&returns, total_position_value)
+    };
     let beta = calculate_beta(market_data, &returns);
     let sharpe_ratio = calculate_sharpe_ratio(&returns);
 
@@ -58,7 +66,12 @@ fn calculate_position_metrics(portfolio: &Portfolio, current_price: Decimal) -> 
     let mut total_cost = dec!(0);
     let mut total_quantity = dec!(0);
 
-    for trade in portfolio.get_open_positions_as_trades().unwrap().iter() {
+    let open_positions = portfolio.get_open_positions_as_trades();
+    if open_positions.is_none() {
+        return (dec!(0), dec!(0), dec!(0))
+    }
+
+    for trade in open_positions.unwrap().iter() {
         let quantity = trade.get_quantity();
         let cost = trade.get_notional_value();
 
@@ -117,6 +130,9 @@ fn calculate_beta(market_data: &[Candle], asset_returns: &[Decimal]) -> Decimal 
 /// This does not account for the risk-free rate, which is a common simplification for algo trading
 /// because it should be negligible for short-term trading.
 fn calculate_sharpe_ratio(returns: &[Decimal]) -> Decimal {
+    if returns.len() < 2 {
+        return dec!(0)
+    }
     let mean_return = returns.iter().sum::<Decimal>() / Decimal::from(returns.len());
     let variance = returns.iter()
         .map(|&r| (r - mean_return) * (r - mean_return))
@@ -144,6 +160,56 @@ fn calculate_returns(candles: &[Candle]) -> Vec<Decimal> {
 }
 
 
+/// Risk metrics for a portfolio
+///
+/// # Measurements
+///
+/// These are the more complex risk metrics returned.
+///
+/// ## Value at Risk (VaR)
+///
+/// Quantifies the level of financial risk over a specific time frame with a given confidence interval.
+///
+/// For example, a 95% confidence interval, means that there is a 5% chance that the portfolio will
+/// lose more than the VaR estimate over the defined period.
+///
+///
+/// ## Beta
+///
+/// Measures the correlation and volatility between the asset and the market.
+///
+/// ### Interpretation
+///
+/// - 1: The asset moves in line with the market.
+/// - > 1: The asset is more volatile than the market.
+/// - < 1: The asset is less volatile than the market.
+/// - = 0: The asset's returns have no correlation with the market.
+/// - Negative: The asset tends to move in the opposite direction of the market.
+///
+///
+/// ## Sharpe Ratio
+///
+/// Measures the additional return for the volatility endured for holding a riskier asset.
+///
+/// ### Interpretation
+///
+/// - > 1: The asset is generating a return above the risk-free rate for the volatility endured.
+/// - < 1: The asset is generating a return below the risk-free rate for the volatility endured.
+///
+/// ### Uses
+///
+/// - **Assessing Strategies:** a strategy with a higher Sharpe ratio is generally considered
+///   better as it provides more return for the same amount of risk.
+/// - **Risk Management:** Aids in understanding if the returns of a strategy justify the risk it's
+///   taking. Crucial for maintaining a balanced risk profile in algorithmic trading.
+/// - **Performance Metric:** Evaluates the performance of algorithms over time.
+/// - **Strategy Optimization:** Can be used as an optimization target, adjusting trading parameters to
+///   maximize the Sharpe ratio, aiming for the best risk-adjusted returns.
+/// - **Capital Allocation:** In a system with multiple trading strategies, the Sharpe ratio can guide capital
+///   allocation/distribution. Strategies with higher Sharpe ratios might receive more capital.
+/// - **Robustness Check:** A consistently high Sharpe ratio across different market conditions can indicate
+/// a robust trading strategy.
+#[derive(Debug, Clone)]
 pub struct PortfolioRisk {
     pub total_position_value: Decimal,
     pub average_entry_price: Decimal,